requirements.txt
streamlit
tensorflow
numpy
opencv-python-headless
pillow
segmentation-models